+++ R/expandTables.R
@@ -5,7 +5,7 @@
 #' \code{expandTables()} generates a data frame and supports two levels.
 #'
 #' @param ... vectors of 2x2 tables
-#' @param exp_name Name of \code{Exposure} Variable
+#' @param exp_name Name of \code{exp} Variable
 #' @param exp_lvl Names of two categories in the order of
 #' Exposed and non-exposed
 #' @param case_name Name of \code{Case} variable
@@ -16,7 +16,7 @@
 #'
 #' \code{expandTables} uses the vectors of \code{2x2} tables and
 #' generates a data frame of at least two columns:
-#' Exposure and Outcome.
+#' exp and case.
 #'
 #' \preformatted{expandTables(c(100, 200, 100, 200))}
 #'
@@ -31,9 +31,9 @@
 #'              strata1 = c(100, 200, 100, 200),
 #'              strata2 = c(100, 200, 100, 200),
 #'              strata3 = c(100, 200, 100, 200),
-#'              exp_name = "Exposure",
+#'              exp_name = "exp",
 #'              exp_lvl = c("exposed", "unexposed"),
-#'              case_name = "Outcome",
+#'              case_name = "case",
 #'              case_lvl = c("case", "control"),
 #'              strata_name = "Strata"
 #' )
@@ -45,10 +45,10 @@
 #' applied.
 #'
 #' \enumerate{
-#'     \item Exposure Name: \code{exposure}
-#'     \item Exposure levels: \code{exposed} and \code{unexposed}
-#'     \item Outcome Name: \code{outcome}
-#'     \item Outcome levels: \code{disease} and \code{healthy}
+#'     \item exp Name: \code{exp}
+#'     \item exp levels: \code{exposed} and \code{unexposed}
+#'     \item case Name: \code{case}
+#'     \item case levels: \code{cases} and \code{controls}
 #'     \item Strata Name: \code{strata}
 #'     \item Note: Strata levels are not considered as vectors must
 #'     be named.
@@ -95,10 +95,10 @@
 #'
 #' @export
 expandTables <- function( ... ,
-                          exp_name = "exposure",
+                          exp_name = "exp",
                           exp_lvl = c("exposed", "unexposed"),
-                          case_name = "outcome",
-                          case_lvl = c("disease", "healthy"),
+                          case_name = "case",
+                          case_lvl = c("cases", "controls"),
                           strata_name = "strata")
 {
     # get vectors within three dots