+++ R/mhor.R
@@ -14,14 +14,12 @@
 #'
 #' @details
 #'
-#'
-#' Value can be set as baseline by specifying `exp_value`. This is used
+#' Rows and Columns can be rearranged by specifying
+#' `exp_value` and `case_value`. This is used
 #' when the exposed and case values are not at the right place in 2x2 tables.
 #'
-#' It produces a table with Odds Ratio,95% CI as well as
-#' p-value. If \code{strata} is specified, `Mantel-Haenzsel` Pooled
-#' estimates of `Odds Ratio` is generated along with Chi-squared test for
-#' homogeneity.
+#' Reference row value can be specified in `exp_value`.
+#'
 #'
 #' Attributable fractions, \code{Attr. Frac. Exp} and  \code{Attr. Frac. Pop}
 #' among exposed and population are calculated when OR is greated than or
@@ -29,7 +27,11 @@
 #' If OR is less than 1, preventable fractions,  \code{Prev. Frac. Exp}
 #' and  \code{Attr. Frac. Pop} are calculated.
 #'
-#' \code{Prop.exp} is proportion among exposed.
+#' It produces a table with Odds Ratio, 95% CI as well as
+#' p-value. If \code{strata} is specified, `Mantel-Haenzsel` Pooled
+#' estimates of `Odds Ratio` is generated along with Chi-squared test for
+#' homogeneity.
+#'
 #'
 #' \strong{Odds Ratio, OR}
 #'
@@ -100,25 +102,6 @@
 #' @examples
 #'
 #' ### Example from Essential Medical Statistics
-#' # Page 160, table 16.5 Hypothetical data from a survey to examine
-#' # the prevalence of asthma among patients at a particular general practice.
-#'
-#' gp <- expandTables(c(81, 995, 57, 867),
-#'                    exp_name = "sex",
-#'                    exp_lvl = c("women", "men"),
-#'                    case_name = "asthma",
-#'                    case_lvl = c("yes", "no"))
-#' gp <- labelData(gp, "Hypothetical data: prevalence of asthma")
-#' gp <- labelVar(gp, sex="Patient's sex", asthma="Having asthma")
-#'
-#' ## check dataset
-#' codebook(gp)
-#'
-#' ## calculate OR
-#' mhor(gp, sex, asthma)
-#'
-#'
-#' ### Example from Essential Medical Statistics
 #' # Page 178, Chapter 18: Controlling for confounding: Stratification
 #' lepto <- expandTables(
 #'     male = c(36, 14, 50, 50), female = c(24, 126, 10, 90),
@@ -144,7 +127,6 @@
 #' mhor(lepto, area, ab, gender, "Rural", "Yes")
 #'
 #'
-#'
 #' @export
 mhor <- function(data, exp , case, strata = NULL, exp_value = NULL,
                  case_value = NULL, rnd = 4)
@@ -161,17 +143,22 @@ mhor <- function(data, exp , case, strata = NULL, exp_value = NULL,
 
     ## if input is not a data.frame, stop
     if (!is.data.frame(.data)) {
-        stop("`.data` must be a data.frame", call. = FALSE)
+        stop(paste0("`", ".data_name", "` must be a data.frame"), call. = FALSE)
     }
-
 
     ## get exp and case name
     exp <- as.character(.args$exp)
     case <- as.character(.args$case)
     strata <- as.character(.args$strata)
 
+    ## check if these variables are in the dataset
+    lapply(c(exp, case) , function(z) {
+        if (!(z %in% .vars_names)) {
+            stop(paste0("'", z, "' not found."), call. = FALSE)
+        }
+    })
 
-    ## create single vector .exp
+    ## create single vectors for exp and case
     .exp <- .data[[exp]]
     .case <- .data[[case]]
 
@@ -200,48 +187,27 @@ mhor <- function(data, exp , case, strata = NULL, exp_value = NULL,
         )
 
         ## create label
-        .lbl <- paste0("      Odds Ratio : '", case, "'")
+        .lbl <- paste0(" Measure of Association : Odds Ratio for '", case, "'")
     } else {
         ## exp var must be binary for MHOR.
         if (nrow(.tbl) > 2) {
             stop(paste0("'", exp, "' must be binary."), call. = FALSE)
         }
 
-        ## get strata levels
-        .lvl <- unique(.data[[strata]])
-        .mhor <- splitStrata(.data, .exp, .case, strata,
-                             .lvl, exp_value, case_value, rnd)
-        ## get crude OR
-        .crude <- calcOR(.tbl, case, rnd)[7, -2]
+        ## create single vector for strata
+        .strata <- .data[[strata]]
+
+        ## calculate MHOR
+        .df <- calcMHOR(.exp, .case, .strata, exp_value, case_value, rnd)
+
+        ## calculate crude OR
+        .crude <- calcOR(table(.exp, .case, useNA = "no"), exp, rnd)[7, -2]
         .crude[1, 1] <- "Crude"
 
-        ## get stratum-specific OR
-        .df <- do.call(
-            rbind,
-            lapply(.lvl, function(z) {
-                .exp <- .exp[.data[[strata]] == z]
-                .case <- .case[.data[[strata]] == z]
-                ## create 2x2 tables
-                .tbl <- table(.exp, .case, useNA = "no")
+        ## add crude OR to MHOR table
+        .df[nrow(.df)-3, ] <- .crude
+        names(.df)[1] <- strata
 
-                ## change row and col orders
-                .tbl <- rowColOrder(.tbl, exp_value, case_value)
-
-                .t <- calcOR(.tbl, case, rnd)[7, -2]
-                .t[1, 1] <- z
-                .t
-            })
-        )
-
-        ## change names to row-combine
-        names(.mhor) <- names(.crude) <- names(.df) <-
-            c(strata, "Odds Ratios", "[95% Conf.", "Interval]", "Pr>chi2")
-        .t <- rbind(.crude, .mhor)
-        .t <- addDashLines(.t)
-        .t <- rbind(.t[-4, ], .t[4, ])
-
-        ## add stratum-specific OR and 95% CI
-        .df <- rbind(.df, .t)
 
         ## add dash lines
         .df <- addDashLines(.df, .vline = 2)
@@ -249,7 +215,6 @@ mhor <- function(data, exp , case, strata = NULL, exp_value = NULL,
         ## create label
         .lbl <- paste0("                Mantel-Haenszel (M-H) Odds Ratio")
     }
-
 
     ## print results
     printDF(.df, .lbl)
@@ -263,20 +228,20 @@ mhor <- function(data, exp , case, strata = NULL, exp_value = NULL,
     invisible(.df)
 }
 
-
 # helpers -----------------------------------------------------------------
 
-calcOR <- function(tbl, x, rnd = 3)
+
+calcOR <- function(.tbl, exp, rnd)
 {
-    ## add rows and then proportions
-    .tbl <- tbl
-    .row_total <- rowSums(.tbl)
-    .col_total <- colSums(cbind(.tbl, .row_total))
-    .tbl_all <- rbind(cbind(.tbl, Total = .row_total), Total = .col_total)
-    .prp_exp <- .tbl_all[ , 1] / .tbl_all[, 3]
+    ## create row and column totals and odds.
+    .tblr <- cbind(.tbl, Total = rowSums(.tbl))
+    .tblc <- rbind(.tblr, Total = colSums(.tblr))
+    .tblf <- cbind(
+        .tblc,
+        Odds = sprintf(.tblc[, 1] / .tblc[, 2], fmt = paste0("%#.", rnd, "f"))
+    )
 
-
-    ## get odds ratios and 95% CI
+    ## calculate odds ratio, 95% CI, and p-value
     .a <- .tbl[1, 1]
     .b <- .tbl[1, 2]
     .c <- .tbl[2, 1]
@@ -295,32 +260,22 @@ calcOR <- function(tbl, x, rnd = 3)
         return(NA)
     })
 
-
     ## calculate attributable or prevent
     .m1 <- .a + .b
     if (.or >= 1) {
         .afe <- (.or - 1) / .or
         .afe_confint <- (.confint - 1) / .confint
         .afp <- .afe * .a / .m1
-        .lbl <- c("Odds ratio" = "Odds Ratio",
-                  "Attributable Fraction among exposed" = "Attr. Frac. Exp",
-                  "Attributable Fraction among population" = "Attr. Frac. Pop")
+        .lbl <- c("Odds Ratio", "Attr. Frac. Exp", "Attr. Frac. Pop")
     } else {
         .afe <- 1 - .or
         .afe_confint <- c(1 - .confint[2], 1- .confint[1])
         .afp <- (.a / .m1 * .afe) / ((.a / .m1 * .afe) + .or)
-        .lbl <- c("Odds ratio" = "Odds Ratio",
-                  "Prevented Fraction among exposed" = "Prev. Frac. Exp",
-                  "Prevented Fraction among population" = "Prev. Frac. Pop")
+        .lbl <- c("Odds Ratio", "Prev. Frac. Exp", "Prev. Frac. Pop")
     }
 
     ## combine all in one dataframe
-    .df <- as.data.frame(.tbl_all)
-    .df <- cbind(row.names(.tbl_all), .df,
-                 "Prop.exp" = sprintf(.prp_exp,
-                                      fmt = paste0("%#.", rnd, "f")))
-    names(.df)[1] <- x
-
+    .df <- data.frame(row.names(.tblf), .tblf)
 
     ## get other estimates
     .est <- data.frame(
@@ -334,14 +289,13 @@ calcOR <- function(tbl, x, rnd = 3)
         c(sprintf(.chi$p.value,
                   fmt = paste0("%#.", rnd, "f")), "", "")
     )
-    names(.est) <- names(.df)
+    names(.est) <- names(.df) <- c(exp, colnames(.tblf))
 
     ## combine into final dataframe
     .df <- rbind(.df,
                  c(" ", "Estimates", "[95% Conf.", "Interval]", "Pr>chi2"),
                  .est)
     row.names(.df) <- NULL
-
 
     ## add dash lines
     .df <- addDashLines(.df)
@@ -353,78 +307,80 @@ calcOR <- function(tbl, x, rnd = 3)
 }
 
 
-splitStrata <- function(.data, .exp, .case, strata, .lvl, exp_value, case_value, rnd)
+calcMHOR <- function(.exp, .case, .strata, exp_value, case_value, rnd)
 {
-    ## calculate MHOR and 95% CI, chi square test for homogeneity
-    ## chi-square test for MHOR
+
+    ## make three tables
+    .tbl <- table(.exp, .case, .strata, useNA = "no")
+    .lvl_len <- dim(.tbl)[3]
+
+    ## calculate MHOR
+    .t <- mantelhaen.test(.tbl, correct = FALSE)
+    .mhor <- c("M-H Combined",
+               sprintf(c(.t$estimate, .t$conf.int, .t$p.value),
+                       fmt = paste0("%#.", rnd, "f")))
+
+    # calculate chi-square and p-value for homogeneity test
+    # Woolf test of homogeneity of odds ratios (Jewell 2004, page 154).
     .t <- do.call(
         rbind,
-        lapply(.lvl, function(z) {
-            .exp <- .exp[.data[[strata]] == z]
-            .case <- .case[.data[[strata]] == z]
-            .t <- table(.exp, .case, useNA = "no")
-
+        lapply(1:.lvl_len, function(z) {
             ## change row and col orders
-            .t <- rowColOrder(.t, exp_value, case_value)
+            .t <- rowColOrder(.tbl[,,z], exp_value, case_value)
+            .strata_name <- dimnames(.tbl)$.strata[z]
+            a <- .t[1, 1]
+            b <- .t[1, 2]
+            c <- .t[2, 1]
+            d <- .t[2, 2]
 
-            calcMHOR(.t)
+            or <- log(((a + 0.5) * (d + 0.5)) /
+                          ((b + 0.5) * (c + 0.5)))
+            var <- (1 / (a + 0.5)) +
+                (1 / (b + 0.5)) + (1 / (c + 0.5)) + (1 / (d + 0.5))
+            w <- 1 / var
+            wor <- w * or
+
+            ## cacluate OR
+            res <- calcOR(.t, "tocombine", rnd)[7, -2]
+            res[1, 1] <- .strata_name
+
+            ## return result
+            .df <- data.frame(res, or, w, wor)
+            names(.df)[1:5] <- c("", "Point Estimate",
+                                 "[95% Conf.", "Interval]", "Pr>chi2")
+            .df
         })
     )
 
-    .mhor <- sum(.t[["q"]]) / sum(.t[["r"]])
-    .mhse <- sqrt(sum(.t[["v"]]) / (sum(.t[["q"]]) * sum(.t[["r"]])))
-    .mhconfint <- c(.mhor / exp(1.96 * .mhse), .mhor * exp(1.96 * .mhse))
 
-    ## test that combine OR = 1
-    .chi <- (sum(.t[["o"]]) - sum(.t[["e"]]))^2 / sum(.t[["v"]])
-    .chi <- tryCatch({
-        suppressWarnings(pchisq(.chi, df = 1, lower.tail = FALSE))
+    lnOR <- sum(.t$wor) / sum(.t$w)
+
+    # Equation 10.3 from Jewell (2004):
+    .chi <- sum(.t$w * (.t$or - lnOR)^2)
+    .pvalue <- tryCatch({
+        suppressWarnings(pchisq(.chi, df = .lvl_len - 1, lower.tail = FALSE))
     }, error = function(err) {
         return(NA)
     })
 
-    ## test of homogeneity
-    .homox <- ((.t[, "o"] * .t[, "h0"]) -
-                   (.mhor * .t[, "d0"] * .t[, "h1"]))^2 /
-        (.mhor * .t[, "v"] * (.t[, "n"]^2))
-    .homox <- sum(.homox)
-    .homo <- tryCatch({
-        suppressWarnings(pchisq(.homox, df = length(.lvl) - 1, lower.tail = FALSE))
-    }, error = function(err) {
-        return(NA)
-    })
 
-    ## return a matrix
-    .df <- rbind(
-        c("M-H combined",
-          sprintf(c(.mhor, .mhconfint, .chi), fmt = paste0('%#.', rnd, 'f'))),
-        c("Homogeneity Test (M-H)",
-          paste0("chi(", length(.lvl) - 1, ")"),
-          sprintf(.homox, fmt = paste0('%#.', 2, 'f')),
-          "Pr>chi2",
-          sprintf(.homo, fmt = paste0('%#.', rnd, 'f')))
+    ## combine estimates
+    .t <- .t[, 1:5]
+    .df <- data.frame(
+        rbind(
+            # c("", "Point Estimate", "[95% Conf.", "Interval]", "Pr>chi2"),
+            c("", "", "", "", ""),
+            .mhor,
+            c("M-H Homogeneity Test", paste0("chi(", .lvl_len - 1, ")"),
+              sprintf(.chi, fmt = paste0('%#.', 2, 'f')), "Pr>chi2",
+              sprintf(.pvalue, fmt = paste0('%#.', rnd, 'f')))
+        )
     )
-    data.frame(.df)
+    names(.df) <- names(.t)
+    .df <- addDashLines(.df)
+    .df <- rbind(.df[-4, ], .df[4, ])
+
+    .df <- rbind(.t, .df)
+    row.names(.df) <- NULL
+    return(.df)
 }
-
-
-calcMHOR <- function(.tbl)
-{
-    d1 <- .tbl[1, 1]
-    h1 <- .tbl[1, 2]
-    d0 <- .tbl[2, 1]
-    h0 <- .tbl[2, 2]
-    d <- d1 + d0
-    h <- h1 + h0
-    n1 <- d1 + h1
-    n0 <- d0 + h0
-    n <- sum(.tbl)
-
-    .q <- d1 * h0 / n
-    .r <- d0 * h1 / n
-    .v <- (d * h * n0 * n1) / ((n)^2 * (n - 1))
-    .e <- (d * n1) / n
-    data.frame(q = .q, r = .r, v = .v, e = .e, o = d1,
-               d0 = d0, h1 = h1, h0 = h0, n = n)
-}
-